requirements.txt
fpdf2
Pillow